+++ apps/sasquatch/src/main/java/com/microsoft/appcenter/sasquatch/listeners/SasquatchCrashesListener.java
@@ -5,20 +5,13 @@
 
 package com.microsoft.appcenter.sasquatch.listeners;
 
-import android.content.ContentResolver;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.database.Cursor;
 import android.net.Uri;
 import android.os.SystemClock;
-import android.provider.OpenableColumns;
 import android.support.annotation.VisibleForTesting;
 import android.support.test.espresso.idling.CountingIdlingResource;
 import android.support.v7.app.AlertDialog;
-import android.text.TextUtils;
-import android.text.format.Formatter;
-import android.util.Log;
-import android.webkit.MimeTypeMap;
 import android.widget.Toast;
 
 import com.microsoft.appcenter.crashes.AbstractCrashesListener;
@@ -26,16 +19,8 @@ import com.microsoft.appcenter.crashes.Crashes;
 import com.microsoft.appcenter.crashes.ingestion.models.ErrorAttachmentLog;
 import com.microsoft.appcenter.crashes.model.ErrorReport;
 import com.microsoft.appcenter.sasquatch.R;
-import com.microsoft.appcenter.sasquatch.activities.MainActivity;
+import com.microsoft.appcenter.sasquatch.util.AttachmentsUtil;
 import com.microsoft.appcenter.utils.HandlerUtils;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.LinkedList;
-import java.util.List;
-
-import static com.microsoft.appcenter.sasquatch.activities.MainActivity.LOG_TAG;
 
 @SuppressWarnings("TryFinallyCanBeTryWithResources")
 public class SasquatchCrashesListener extends AbstractCrashesListener {
@@ -47,10 +32,6 @@ public class SasquatchCrashesListener extends AbstractCrashesListener {
 
     private static final long TOAST_DELAY = 2000;
 
-    private String mTextAttachment;
-
-    private Uri mFileAttachment;
-
     private long mBeforeSendingToastTime;
 
     public SasquatchCrashesListener(Context context) {
@@ -58,19 +39,19 @@ public class SasquatchCrashesListener extends AbstractCrashesListener {
     }
 
     public String getTextAttachment() {
-        return mTextAttachment;
+        return AttachmentsUtil.getInstance().getTextAttachment();
     }
 
     public void setTextAttachment(String textAttachment) {
-        this.mTextAttachment = textAttachment;
+        AttachmentsUtil.getInstance().setTextAttachment(textAttachment);
     }
 
     public Uri getFileAttachment() {
-        return mFileAttachment;
+        return AttachmentsUtil.getInstance().getFileAttachment();
     }
 
     public void setFileAttachment(Uri fileAttachment) {
-        this.mFileAttachment = fileAttachment;
+        AttachmentsUtil.getInstance().setFileAttachment(fileAttachment);
     }
 
     @Override
@@ -103,32 +84,9 @@ public class SasquatchCrashesListener extends AbstractCrashesListener {
 
     @Override
     public Iterable<ErrorAttachmentLog> getErrorAttachments(ErrorReport report) {
-        List<ErrorAttachmentLog> attachments = new LinkedList<>();
-
-        /* Attach app icon to test binary. */
-        if (mFileAttachment != null) {
-            try {
-                byte[] data = getFileAttachmentData();
-                String name = getFileAttachmentDisplayName();
-                String mime = getFileAttachmentMimeType();
-                ErrorAttachmentLog binaryLog = ErrorAttachmentLog.attachmentWithBinary(data, name, mime);
-                attachments.add(binaryLog);
-            } catch (SecurityException e) {
-                Log.e(LOG_TAG, "Couldn't get file attachment data.", e);
-
-                /* Reset file attachment. */
-                MainActivity.setFileAttachment(null);
-            }
-        }
-
-        /* Attach some text. */
-        if (!TextUtils.isEmpty(mTextAttachment)) {
-            ErrorAttachmentLog textLog = ErrorAttachmentLog.attachmentWithText(mTextAttachment, "text.txt");
-            attachments.add(textLog);
-        }
 
         /* Return attachments as list. */
-        return attachments.size() > 0 ? attachments : null;
+        return AttachmentsUtil.getInstance().getErrorAttachments(mContext);
     }
 
     @Override
@@ -166,83 +124,11 @@ public class SasquatchCrashesListener extends AbstractCrashesListener {
         crashesIdlingResource.decrement();
     }
 
-    public String getFileAttachmentDisplayName() throws SecurityException {
-        Cursor cursor = mContext.getContentResolver()
-                .query(mFileAttachment, null, null, null, null);
-        try {
-            if (cursor != null && cursor.moveToFirst()) {
-                int nameIndex = cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME);
-                if (!cursor.isNull(nameIndex)) {
-                    return cursor.getString(nameIndex);
-                }
-            }
-        } finally {
-            if (cursor != null) {
-                cursor.close();
-            }
-        }
-        return "";
+    public String getFileAttachmentDisplayName() {
+        return AttachmentsUtil.getInstance().getFileAttachmentDisplayName(mContext);
     }
 
     public String getFileAttachmentSize() throws SecurityException {
-        Cursor cursor = mContext.getContentResolver()
-                .query(mFileAttachment, null, null, null, null);
-        try {
-            if (cursor != null && cursor.moveToFirst()) {
-                int sizeIndex = cursor.getColumnIndex(OpenableColumns.SIZE);
-                if (!cursor.isNull(sizeIndex)) {
-                    return Formatter.formatFileSize(mContext, cursor.getLong(sizeIndex));
-                }
-            }
-        } finally {
-            if (cursor != null) {
-                cursor.close();
-            }
-        }
-        return "Unknown";
-    }
-
-    private byte[] getFileAttachmentData() throws SecurityException {
-        InputStream inputStream = null;
-        ByteArrayOutputStream outputStream = null;
-        try {
-            inputStream = mContext.getContentResolver().openInputStream(mFileAttachment);
-            if (inputStream == null) {
-                return null;
-            }
-            outputStream = new ByteArrayOutputStream();
-            byte[] buffer = new byte[4096];
-            int read;
-            while ((read = inputStream.read(buffer)) != -1) {
-                outputStream.write(buffer, 0, read);
-            }
-        } catch (IOException e) {
-            Log.e(LOG_TAG, "Couldn't read file", e);
-        } finally {
-            try {
-                if (inputStream != null) {
-                    inputStream.close();
-                }
-            } catch (IOException ignore) {
-            }
-            try {
-                if (outputStream != null) {
-                    outputStream.close();
-                }
-            } catch (IOException ignore) {
-            }
-        }
-        return outputStream != null ? outputStream.toByteArray() : null;
-    }
-
-    private String getFileAttachmentMimeType() {
-        String mimeType;
-        if (ContentResolver.SCHEME_CONTENT.equals(mFileAttachment.getScheme())) {
-            mimeType = mContext.getContentResolver().getType(mFileAttachment);
-        } else {
-            String fileExtension = MimeTypeMap.getFileExtensionFromUrl(mFileAttachment.toString());
-            mimeType = MimeTypeMap.getSingleton().getMimeTypeFromExtension(fileExtension.toLowerCase());
-        }
-        return mimeType;
+        return AttachmentsUtil.getInstance().getFileAttachmentSize(mContext);
     }
 }